+++ src/schemathesis/specs/openapi/stateful/links.py
@@ -54,21 +54,21 @@ class OpenApiLink:
     )
 
     def __init__(self, name: str, status_code: str, definition: dict[str, Any], source: APIOperation):
-        from schemathesis.specs.openapi.schemas import BaseOpenAPISchema
+        from schemathesis.specs.openapi.schemas import OpenApiSchema
 
         self.name = name
         self.status_code = status_code
         self.source = source
-        assert isinstance(source.schema, BaseOpenAPISchema)
+        assert isinstance(source.schema, OpenApiSchema)
         errors = []
 
         get_operation: Callable[[str], APIOperation]
         if "operationId" in definition:
             operation_reference = definition["operationId"]
-            get_operation = source.schema.get_operation_by_id
+            get_operation = source.schema.find_operation_by_id
         else:
             operation_reference = definition["operationRef"]
-            get_operation = source.schema.get_operation_by_reference
+            get_operation = source.schema.find_operation_by_reference
 
         try:
             self.target = get_operation(operation_reference)